Hey,

two things:

1. Performance
The `TGAImageReader` parser didn't buffer image data during reading. I gained a speedup of factor 40 on my machine by simply reading the pixel data with `readFully(..)` instead of `readByte()` or `readShort()` (~~since buffer inside `javax.imageio.stream.ImageInputStreamImpl` gets activated~~ Edit: I actually implemented a new buffer inside `ByteBuffer` in this PR) .

2. OpenCV
I recently tried to integrate JavaCV / OpenCV into my project. It's a native C++ library that has some nice & fast image functionality. When integrating it i noticed that the internal pixel format that the TGAImageReader creates is not optima for OpenCV, because if seems to expect BGR(A) `int`, but it gets RGB(A) `int`.  
Anyway, i changed the format to BGR(A) `byte[]` (interleaved instead of packed) which can be read by OpenCV. But i did not test these changes with non-24/32 bit per channel files, so if maybe sombody else could look over these changes maybe even test them that would be awesome.
If this second change is not desirable let me know.