@Katharsas 2018-09-28
Hmm.
Actually i buffered the entire file it seems, that is not what i wanted. Let me fix that.