@tmyroadctfig 2018-10-08
The change looks good to me aside from that one query about the order of the `ByteBuffer` calls. I'll see if I can find some other test data to check those other paths.