@Katharsas 2018-10-08
I see now that my IDE replaced spaces with tabs. I could revert that too to make the diff cleaner.