@Katharsas 2018-12-02
Ready for merge, still more testing wanted. I can squash if needed.