@Katharsas 2018-12-03
Hey!
I did some more performance testing because i was suspecting an error in the readme (the part where it warns from using InputStreams), but apparently the readme is correct. Here are the times from the microbenchmark i did on a single 32 bit 1024*1024 image.

Machine:
CPU: i7 7700HQ, SSD: Samsung SSD 840 Evo 500GB

**Before this pull request**
(average of 5 samples per method after warmup):
- `ImageIO.read(path.toFile());`
  2569.6 ms
- `ImageIO.read(new FileInputStream(path.toFile()));`
9315.8 ms
- `ImageIO.read(new BufferedInputStream(new FileInputStream(path.toFile())));`
  6834.6 ms

**With this pull request**
(average of 100 samples per method after warmup):
- `ImageIO.read(path.toFile());`
  10.35 ms
- `ImageIO.read(new FileInputStream(path.toFile()));`
38.05 ms
- `ImageIO.read(new BufferedInputStream(new FileInputStream(path.toFile())));`
  32.25 ms

So it seems the speedups are even larger than i thought, this is factor x250  :)
And the readme is right, first method is fastest always.