@Katharsas 2018-12-10
My buffer doesn't work with RLE. Fixed in https://github.com/tmyroadctfig/com.realityinteractive.imageio.tga/pull/7